Files for calibrating WheeTrometer spectrometer
A simple user interface for writing data to the non-volatile EEPROM memory of a WheeTrometer spectrometer.  The interface can be used to assign a serial number to a spectrometer.  It can be used to acquire and save spectral data to a host computer.  It can be used to read saved data.  The data can be used to optimize amplification and offset parameters and these parameters can be written to the EEPROM of the spectrometer.  Wavelength calibration can be done from spectra with known emission max.  The emission spectra of a fluorescent lamp is a convenient example that is built into the software.  Using this data, calibration parameters can be determined by least squares refinement.  The parameters can be written to EEPROM, and will later be read back to the user interface.